scripts/test.py
# -*- coding: utf-8 -*-
"""
Script de testes automatizados para o sistema distribuído
"""
import subprocess
import time
import json
import zmq
import msgpack
import sys
import os

# Configura encoding para Windows
if sys.platform == 'win32':
    os.environ['PYTHONIOENCODING'] = 'utf-8'

def send_msgpack(sock, obj):
    data = msgpack.packb(obj, use_bin_type=True)
    sock.send(data)

def recv_msgpack(sock):
    data = sock.recv()
    return msgpack.unpackb(data, raw=False)

def test_reference_service():
    """Testa o serviço de referência"""
    print("\n[TESTE] Testando serviço de referência...")
    try:
        ctx = zmq.Context()
        req = ctx.socket(zmq.REQ)
        req.setsockopt(zmq.LINGER, 0)
        req.setsockopt(zmq.RCVTIMEO, 5000)
        req.connect("tcp://localhost:5559")
        
        # Testa serviço de rank
        msg = {
            "service": "rank",
            "data": {
                "user": "test_server",
                "timestamp": time.time(),
            },
        }
        send_msgpack(req, msg)
        resp = recv_msgpack(req)
        req.close()
        ctx.term()
        
        if resp.get("service") == "rank" and resp.get("data", {}).get("rank"):
            print("   ✅ Serviço de referência respondendo")
            return True
        else:
            print(f"   ❌ Resposta inesperada: {resp}")
            return False
    except Exception as e:
        print(f"   ❌ Erro ao conectar: {e}")
        return False

def test_server_connection(server_name="server_1"):
    """Testa se o servidor está rodando e respondendo"""
    print(f"\n[TESTE] Testando servidor {server_name}...")
    try:
        # Verifica se o container está rodando
        result = subprocess.run(
            ["docker", "compose", "ps", server_name],
            capture_output=True,
            text=True,
            timeout=5
        )
        if "Up" in result.stdout:
            print(f"   ✅ {server_name} está rodando")
            # Verifica logs para ver se está respondendo
            log_result = subprocess.run(
                ["docker", "compose", "logs", "--tail", "10", server_name],
                capture_output=True,
                text=True,
                timeout=5
            )
            logs = log_result.stdout.lower()
            if "online" in logs or "rep em" in logs or "conectado" in logs:
                print(f"   ✅ {server_name} parece estar funcionando (verificado nos logs)")
                return True
            else:
                print(f"   ⚠️  {server_name} está rodando mas logs podem não mostrar atividade ainda")
                return True  # Container está rodando, assumimos que está OK
        else:
            print(f"   ❌ {server_name} não está rodando")
            return False
    except Exception as e:
        print(f"   ❌ Erro ao verificar {server_name}: {e}")
        return False

def test_bots_running():
    """Verifica se os bots estão rodando"""
    print("\n[TESTE] Verificando se bots estão rodando...")
    try:
        result = subprocess.run(
            ["docker", "compose", "ps", "bot_1", "bot_2"],
            capture_output=True,
            text=True,
            timeout=10
        )
        if "Up" in result.stdout:
            print("   ✅ Bots estão rodando")
            return True
        else:
            print("   ❌ Bots não estão rodando")
            print(f"   Output: {result.stdout}")
            return False
    except Exception as e:
        print(f"   ❌ Erro ao verificar bots: {e}")
        return False

def test_bot_messages():
    """Verifica se os bots estão enviando mensagens"""
    print("\n[TESTE] Verificando mensagens dos bots...")
    try:
        # Verifica logs dos bots
        result = subprocess.run(
            ["docker", "compose", "logs", "--tail", "20", "bot_1", "bot_2"],
            capture_output=True,
            text=True,
            timeout=10
        )
        logs = result.stdout.lower()
        if "mensagem enviada" in logs or "msg" in logs or "publish" in logs:
            print("   ✅ Bots estão enviando mensagens")
            return True
        else:
            print("   ⚠️  Nenhuma mensagem encontrada nos logs recentes")
            print("   (Isso pode ser normal se os bots acabaram de iniciar)")
            return True  # Não é um erro crítico
    except Exception as e:
        print(f"   ❌ Erro ao verificar logs: {e}")
        return False

def test_election():
    """Verifica se a eleição de coordenador está funcionando"""
    print("\n[TESTE] Verificando eleição de coordenador...")
    try:
        # Verifica logs dos servidores
        result = subprocess.run(
            ["docker", "compose", "logs", "--tail", "30", "server_1", "server_2", "server_3"],
            capture_output=True,
            text=True,
            timeout=10
        )
        logs = result.stdout.lower()
        if "coordenador" in logs or "coordinator" in logs or "eleito" in logs:
            print("   ✅ Eleição de coordenador encontrada nos logs")
            return True
        else:
            print("   ⚠️  Nenhuma menção de coordenador nos logs")
            print("   Verificando se servidores têm rank...")
            # Verifica se servidores têm rank
            if "rank recebido" in logs or "rank" in logs:
                print("   ✅ Servidores têm rank")
                return True
            else:
                print("   ❌ Servidores podem não ter rank ainda")
                return False
    except Exception as e:
        print(f"   ❌ Erro ao verificar logs: {e}")
        return False

def test_channels():
    """Testa criação e listagem de canais"""
    print("\n[TESTE] Testando criação e listagem de canais...")
    # Verifica através dos logs se há canais
    try:
        result = subprocess.run(
            ["docker", "compose", "logs", "--tail", "50", "server_1"],
            capture_output=True,
            text=True,
            timeout=10
        )
        logs = result.stdout.lower()
        # Verifica se há menção de canais nos logs
        if "canal" in logs or "channel" in logs:
            print("   ✅ Canais encontrados nos logs")
            return True
        else:
            # Verifica se há mensagens (que indicam que há canais)
            if "publish" in logs or "msg" in logs:
                print("   ✅ Mensagens encontradas (indicando canais ativos)")
                return True
            else:
                print("   ⚠️  Nenhum canal encontrado nos logs (pode ser normal se o sistema acabou de iniciar)")
                return True  # Não é um erro crítico
    except Exception as e:
        print(f"   ❌ Erro ao verificar canais: {e}")
        return False

def test_logical_clock():
    """Testa se o relógio lógico está funcionando"""
    print("\n[TESTE] Testando relógio lógico...")
    # Verifica através dos logs se o relógio lógico está sendo usado
    try:
        result = subprocess.run(
            ["docker", "compose", "logs", "--tail", "50", "server_1", "bot_1"],
            capture_output=True,
            text=True,
            timeout=10
        )
        logs = result.stdout.lower()
        # Verifica se há menção de clock nos logs
        if "clock" in logs:
            print("   ✅ Relógio lógico encontrado nos logs")
            return True
        else:
            print("   ⚠️  Nenhuma menção de relógio nos logs (pode ser normal)")
            return True  # Não é crítico, pois o relógio pode estar funcionando sem logs
    except Exception as e:
        print(f"   ❌ Erro ao verificar relógio: {e}")
        return False

def test_servers_status():
    """Verifica status dos servidores"""
    print("\n[TESTE] Verificando status dos servidores...")
    try:
        result = subprocess.run(
            ["docker", "compose", "ps", "server_1", "server_2", "server_3"],
            capture_output=True,
            text=True,
            timeout=10
        )
        output = result.stdout
        servers_up = output.count("Up")
        if servers_up >= 3:
            print(f"   ✅ {servers_up} servidor(es) rodando")
            return True
        else:
            print(f"   ⚠️  Apenas {servers_up} servidor(es) rodando (esperado: 3)")
            return servers_up > 0
    except Exception as e:
        print(f"   ❌ Erro ao verificar servidores: {e}")
        return False

def main():
    print("=" * 70)
    print("TESTES AUTOMATIZADOS DO SISTEMA DISTRIBUÍDO")
    print("=" * 70)
    
    results = {}
    
    # Aguarda um pouco para o sistema estabilizar
    print("\n[Aguardando] Aguardando 5 segundos para o sistema estabilizar...")
    time.sleep(5)
    
    # Executa testes
    results["reference"] = test_reference_service()
    results["servers_status"] = test_servers_status()
    results["server_connection"] = test_server_connection()
    results["election"] = test_election()
    results["bots_running"] = test_bots_running()
    results["bot_messages"] = test_bot_messages()
    results["channels"] = test_channels()
    results["logical_clock"] = test_logical_clock()
    
    # Resumo
    print("\n" + "=" * 70)
    print("RESUMO DOS TESTES")
    print("=" * 70)
    
    passed = sum(1 for v in results.values() if v)
    total = len(results)
    
    for test_name, result in results.items():
        status = "✅ PASSOU" if result else "❌ FALHOU"
        print(f"   {test_name:20s} : {status}")
    
    print(f"\n   Total: {passed}/{total} testes passaram")
    
    if passed == total:
        print("\n   🎉 Todos os testes passaram!")
        return 0
    else:
        print(f"\n   ⚠️  {total - passed} teste(s) falharam")
        return 1

if __name__ == "__main__":
    sys.exit(main())

